RWE-14 Password change not reflected in auth server
Regardless of the failure mode on authentication with the auth server, we'd
try local authentication.  So, if the password was wrong, we'd fall back to local.
Changed so that we only try local if there is no auth_service entry for the
roxiware server.

Also, password change dialog wasn't authenticating to to auth server when
submitting data.  Changed so that it submits to local server, and the local
server proxies to the auth server