RWE-101 When using an expired password reset link, bad error message
https://roxiware.atlassian.net/browse/RWE-101

Catch the devise-generated error message and replace with a more readable one.

Also, routing url helper generation in 'actionpack' now assumes a 'controller' function/attribute
be available on the caller.  Doesn't care if it has a value, just that it exists.
Default mailer doesn't have one, so add a dummy controller attribute accessor